Finish refactor (#7383)
This PR finally refactors the townhall UI. It seperates each of the pages into its own Window with the following benefits:
a) Better separation of concerns, it's easier to extend existing pages and easier to find certain functionalities.
b) It cleans up a lot of the hacky page management resulting from the double navigation (tabs + pages).
