+++ src/main/java/com/minecolonies/coremod/colony/buildings/workerbuildings/BuildingTownHall.java
@@ -14,7 +14,8 @@ import com.minecolonies.api.colony.colonyEvents.registry.ColonyEventDescriptionT
 import com.minecolonies.api.colony.permissions.PermissionEvent;
 import com.minecolonies.api.util.Log;
 import com.minecolonies.coremod.MineColonies;
-import com.minecolonies.coremod.client.gui.townhall.WindowTownHall;
+import com.minecolonies.coremod.client.gui.townhall.WindowInfoPage;
+import com.minecolonies.coremod.client.gui.townhall.WindowMainPage;
 import com.minecolonies.coremod.colony.buildings.AbstractBuilding;
 import com.minecolonies.coremod.colony.buildings.views.AbstractBuildingView;
 import net.minecraft.network.PacketBuffer;
@@ -200,7 +201,7 @@ public class BuildingTownHall extends AbstractBuilding implements ITownHall
         @Override
         public Window getWindow()
         {
-            return new WindowTownHall(this);
+            return new WindowMainPage(this);
         }
 
         @Override
