+++ src/main/resources/assets/minecolonies/gui/townhall/layoutactions.xml
@@ -1,40 +1,36 @@
-<window xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" size="374 243" pause="false" lightbox="false"
-        xsi:noNamespaceSchemaLocation="file:../../../../java/com/minecolonies/blockout/blockOut.xsd">
+<window size="524 243" pause="false" lightbox="false">
 
-    <switch id="switchpageActions" size="100% 100%">
-        <view id="pageActions">
-            <label size="148 11" color="green" pos="25 24"
-                   label="$(com.minecolonies.coremod.gui.townHall.currTownHallName)"/>
-            <label size="148 11" id="name" color="black" pos="25 35"/>
-            <label size="135 11" id="level" color="black" pos="38 69"/>
+    <layout source="minecolonies:gui/townhall/windowtownhall.xml"/>
+    <label size="148 11" color="green" pos="100 24"
+           label="$(com.minecolonies.coremod.gui.townHall.currTownHallName)"/>
+    <label size="148 11" id="name" color="black" pos="100 35"/>
+    <label size="135 11" id="level" color="black" pos="113 69"/>
 
-            <button id="build" size="120 19" pos="38 107"
-                    label="$(com.minecolonies.coremod.gui.townHall.build)"/>
-<!--        <button id="changeSpec" size="120 19" pos="38 153"
-                    label="$(com.minecolonies.coremod.gui.townHall.togglespec)"/> -->
-            <button id="rename" size="120 19" pos="38 130"
-                    label="$(com.minecolonies.coremod.gui.townHall.rename)"/>
-            <button id="mercenaries" size="120 19" pos="38 153"
-                    label="$(com.minecolonies.coremod.gui.townHall.mercenary)"/>
+    <button id="build" size="120 19" pos="113 107"
+            label="$(com.minecolonies.coremod.gui.townHall.build)"/>
+    <!--        <button id="changeSpec" size="120 19" pos="38 153"
+                        label="$(com.minecolonies.coremod.gui.townHall.togglespec)"/> -->
+    <button id="rename" size="120 19" pos="113 130"
+            label="$(com.minecolonies.coremod.gui.townHall.rename)"/>
+    <button id="mercenaries" size="120 19" pos="113 153"
+            label="$(com.minecolonies.coremod.gui.townHall.mercenary)"/>
 
-            <label size="148 11" pos="201 30" style="bold" color="green"
-                   label="$(com.minecolonies.coremod.gui.townHall.allies)"/>
-            <list id="allies" size="150 75" pos="201 40">
-                <box size="100% 25" linewidth="2">
-                    <label id="name" size="100 12" pos="5 0" color="black"/>
-                    <label id="dist" size="40 13" pos="112 12" color="black"/>
-                    <button id="tp" size="100 12" pos="5 12" label="$(com.minecolonies.coremod.permission.teleport_to_colony)"/>
-                </box>
-            </list>
+    <label size="148 11" pos="271 30" style="bold" color="green"
+           label="$(com.minecolonies.coremod.gui.townHall.allies)"/>
+    <list id="allies" size="150 75" pos="271 40">
+        <box size="100% 25" linewidth="2">
+            <label id="name" size="100 12" pos="5 0" color="black"/>
+            <label id="dist" size="40 13" pos="112 12" color="black"/>
+            <button id="tp" size="100 12" pos="5 12" label="$(com.minecolonies.coremod.permission.teleport_to_colony)"/>
+        </box>
+    </list>
 
-            <label size="148 11" pos="201 130" style="bold" color="red"
-                   label="$(com.minecolonies.coremod.gui.townHall.feuds)"/>
-            <list id="feuds" size="150 50" pos="205 140">
-                <box size="100% 25" linewidth="2">
-                    <label id="name" size="100 12" pos="5 0" color="black"/>
-                    <label id="dist" size="40 12" pos="105 0" color="black"/>
-                </box>
-            </list>
-        </view>
-    </switch>
+    <label size="148 11" pos="271 130" style="bold" color="red"
+           label="$(com.minecolonies.coremod.gui.townHall.feuds)"/>
+    <list id="feuds" size="150 50" pos="271 140">
+        <box size="100% 25" linewidth="2">
+            <label id="name" size="100 12" pos="5 0" color="black"/>
+            <label id="dist" size="40 12" pos="105 0" color="black"/>
+        </box>
+    </list>
 </window>
