+++ src/main/resources/assets/minecolonies/gui/townhall/layoutsettings.xml
@@ -1,39 +1,35 @@
-<window xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" size="374 243" pause="false" lightbox="false"
-        xsi:noNamespaceSchemaLocation="file:../../../../java/com/minecolonies/blockout/blockOut.xsd">
+<window size="524 243" pause="false" lightbox="false">
 
-    <switch id="switchpageSettings" size="100% 100%">
-        <view>
-            <label id="assignCitizens" label="$(com.minecolonies.coremod.gui.townHall.jobHiring)" pos="25 62" size="148 11" textalign="MIDDLE" color="black"/>
-            <button id="toggleJob" size="57 20" pos="70 75" label="$(com.minecolonies.coremod.gui.hiring.off)"/>
+    <layout source="minecolonies:gui/townhall/windowtownhall.xml"/>
+    <label id="assignCitizens" label="$(com.minecolonies.coremod.gui.townHall.jobHiring)" pos="100 62" size="148 11" textalign="MIDDLE" color="black"/>
+    <button id="toggleJob" size="57 20" pos="145 75" label="$(com.minecolonies.coremod.gui.hiring.off)"/>
 
-            <label id="assignHouses" label="$(com.minecolonies.coremod.gui.townHall.housing)" pos="25 102" size="148 11" textalign="MIDDLE" color="black"/>
-            <button id="toggleHousing" size="57 20" pos="70 115" label="$(com.minecolonies.coremod.gui.hiring.off)"/>
+    <label id="assignHouses" label="$(com.minecolonies.coremod.gui.townHall.housing)" pos="100 102" size="148 11" textalign="MIDDLE" color="black"/>
+    <button id="toggleHousing" size="57 20" pos="145 115" label="$(com.minecolonies.coremod.gui.hiring.off)"/>
 
-            <label id="movingIn" label="$(com.minecolonies.coremod.gui.townHall.movein)" pos="25 182" size="148 11" textalign="MIDDLE" color="black"/>
-            <button id="toggleMoveIn" size="57 20" pos="70 195" label="$(com.minecolonies.coremod.gui.townHall.off)"/>
+    <label id="movingIn" label="$(com.minecolonies.coremod.gui.townHall.movein)" pos="100 182" size="148 11" textalign="MIDDLE" color="black"/>
+    <button id="toggleMoveIn" size="57 20" pos="145 195" label="$(com.minecolonies.coremod.gui.townHall.off)"/>
 
-            <label id="printProgress" label="$(com.minecolonies.coremod.gui.townHall.printProgress)" pos="25 142" size="148 11" textalign="MIDDLE" color="black"/>
-            <button id="togglePrintProgress" size="57 20" pos="70 155" label="$(com.minecolonies.coremod.gui.townHall.on)"/>
+    <label id="printProgress" label="$(com.minecolonies.coremod.gui.townHall.printProgress)" pos="100 142" size="148 11" textalign="MIDDLE" color="black"/>
+    <button id="togglePrintProgress" size="57 20" pos="145 155" label="$(com.minecolonies.coremod.gui.townHall.on)"/>
 
-            <label id="pickColor" label="$(com.minecolonies.coremod.gui.townHall.pickTeamColor)" pos="201 28" size="148 11" textalign="MIDDLE" color="black"/>
-            <buttonimage id="previousColor" size="14 15" pos="217 42" label="&lt;" source="minecolonies:textures/gui/builderhut/builder_button_mini.png" textcolor="black"/>
-            <dropdown id="colorPicker"
-                      size="86 17"
-                      pos="231 42"
-                      maxContentHeight="75"
-                      dropDownSize="86 75"
-                      source="minecolonies:textures/gui/builderhut/builder_button_medium.png"
-                      textcolor="black"
-                      scrollbarOffset="9 0"
-                      dropfixx="75">
-                <view size="100% 15">
-                    <label id="id" size="100% 100%" label="" pos="0 0" textalign="MIDDLE"/>
-                    <buttonimage id="button" size="86 17" pos="0 0" source="minecolonies:textures/gui/builderhut/builder_button_medium.png" textcolor="black"/>
-                </view>
-            </dropdown>
-            <buttonimage id="nextColor" size="14 15" pos="317 42" label="&gt;" source="minecolonies:textures/gui/builderhut/builder_button_mini.png" textcolor="black"/>
-
-            <button id="bannerPicker" label="$(com.minecolonies.coremod.gui.townhall.edit_flag)" pos="217 70" size="114 20" />
+    <label id="pickColor" label="$(com.minecolonies.coremod.gui.townHall.pickTeamColor)" pos="276 28" size="148 11" textalign="MIDDLE" color="black"/>
+    <buttonimage id="previousColor" size="14 15" pos="292 42" label="&lt;" source="minecolonies:textures/gui/builderhut/builder_button_mini.png" textcolor="black"/>
+    <dropdown id="colorPicker"
+              size="86 17"
+              pos="306 42"
+              maxContentHeight="75"
+              dropDownSize="86 75"
+              source="minecolonies:textures/gui/builderhut/builder_button_medium.png"
+              textcolor="black"
+              scrollbarOffset="9 0"
+              dropfixx="75">
+        <view size="100% 15">
+            <label id="id" size="100% 100%" label="" pos="0 0" textalign="MIDDLE"/>
+            <buttonimage id="button" size="86 17" pos="0 0" source="minecolonies:textures/gui/builderhut/builder_button_medium.png" textcolor="black"/>
         </view>
-    </switch>
+    </dropdown>
+    <buttonimage id="nextColor" size="14 15" pos="392 42" label="&gt;" source="minecolonies:textures/gui/builderhut/builder_button_mini.png" textcolor="black"/>
+
+    <button id="bannerPicker" label="$(com.minecolonies.coremod.gui.townhall.edit_flag)" pos="292 70" size="114 20"/>
 </window>
